requirements.txt
aioredis==1.1.0
amqp==2.3.2
asgiref==2.3.2
asn1crypto==0.24.0
async-timeout==3.0.0
attrs==18.1.0
autobahn==18.6.1
Automat==0.7.0
Babel==2.6.0
billiard==3.5.0.4
cached-property==1.5.1
celery==4.2.1
certifi==2018.8.24
cffi==1.11.5
channels==2.1.6
channels-redis==2.2.1
chardet==3.0.4
colorama==0.3.9
constantly==15.1.0
cryptography==2.4.2
daphne==2.2.1
Django==2.1.5
django-appconf==1.0.2
django-compressor==2.2
django-filter==2.0.0
django-guardian==1.4.9
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
dnspython==1.16.0
drf-nested-routers==0.90.2
eventlet==0.24.1
flower==0.9.2
greenlet==0.4.15
hiredis==0.2.0
httpie==0.9.9
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
jsonfield==2.0.2
kombu==4.2.1
librabbitmq==2.0.0
monotonic==1.5
msgpack==0.5.6
msgpack-python==0.5.6
mysqlclient==1.3.13
ndg-httpsclient==0.5.1
netifaces==0.10.7
pika==0.11.2
Pillow==5.2.0
ply==3.11
pyasn1==0.4.5
pyasn1-modules==0.2.3
pycparser==2.19
pycryptodome==3.7.2
Pygments==2.2.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pyOpenSSL==18.0.0
pypiwin32==223
python-http-client==3.1.0
pytz==2018.5
pywin32==223
razorpay==1.1.1
rcssmin==1.0.6
redis==2.10.6
requests==2.21.0
rjsmin==1.0.12
sendgrid==5.6.0
service-identity==18.1.0
six==1.11.0
slimit==0.8.1
sorl-thumbnail==12.5.0
tornado==5.1.1
Twisted==18.7.0
txaio==2.10.0
urllib3==1.23
vine==1.1.4
websocket-client==0.54.0
zope.interface==4.5.0
